Refactor: Move BitbakeDriver to common library
In order to share this between the language client and server, it must
be in the common library. Because vscode cannot be imported in the
server, the settings must be passed in from the client. We use the any
type as this is what's produced by vscode.workspace.getConfiguration in
both the client and server.

Tests and references are also updated.