Feat: Have broader bitbake settings sanity checks
When the user opens a new workspace he'll probably need to configure the
extension. Sanity checks are here to check that the settings are
remotely accurate. They don't check that bitbake is fully functionnal
but quickly allow detecting settings errors.

Previously, the only check and notification regarded the bitbake folder.
Now, the environment script is also checked. The build path doesn't need
to be checked because it is dynamically created.

The notification message has been changed to be more generic.
The "Open Settings" button now opens the workspace settings with a more
accurate filter.