Refactor: Use parse task from the server
The language server had a duplicated implementation of the parse task
and problems matching. Through a request, use the one from the client.

Launching the task returns the child execution, but not
the command output, so we don't have access to the return value anymore.
This doesn't cause any errors and actually allows to scan the project
even if some recipes are broken. We now will keep access to the list of
recipes and overrides in those cases.