Refactor: Move shell sanitization to BitbakeSettings
The sanitization should be done by the classes that decide if the string
comes from the user or not. This also calls the function less
frequently.

We'll need to run more complex scripts that need bash features for the
bitbake server, so we'll need to be able to pass in more complex
commands.

In this commit, the command is not sanitized because it's always
provided programmatically by the extension without a risk for user
shell injection. THis follows the best practices documented by VSCode.