Fix: Use a dark theme for the marketplace banner and logo
The web marketplace uses a light theme by default. The logo was not
visible. We opt for a circle outline corresponding to the default VSCode
dark brackground. It is now visible in the marketplace either using
white or black theme.

The .png source has been added (.svg) because it's no longer just an
export from the official logo.