Fix: use workspaceFolder var for python extraPaths
If we set the python extraPaths to the expanded pathToBitbakeFolder+'/lib', it will
result in an absolute path, i.e.

    "python.autoComplete.extraPaths": [
        "/home/patrick/git/<workspace>/poky/bitbake/lib"
    ],

Settings with such an absolute path cannot be version controlled.
Instead, use the path before it is expanded and still contains the
$workspaceFolder variable. Then, the written settings look like this:

    "python.autoComplete.extraPaths": [
        "${workspaceFolder}/poky/bitbake/lib"
    ],

This is more portable and can be tracked via git.