Fix: Stop saving all open editors when saving a bitbake file
The default behavior of vscode is to save all open editors when a task
is run. We currently run the parsing task for any bitbake file save.
This means that if you have multiple bitbake files open, all of them
will be saved when any one of them is saved which is not the desired
behavior.

This patch temporarily disables the saveBeforeRun setting when the
bitbake/parseAllRecipes request is received.