Refactor: Extract BitbakeTerminal class from BitbakeTaskProvider
This class will be used to run bitbake commands in a dedicated terminal
rather than in the background or in the task terminal.

This will allow us to have a better control over displaying the output
of bitbake and exit codes.

Another possibility would be to create a dedicated Output channel for
bitbake but it would be more hidden, and also cannot display queued
commands.