Fix: devtool-open-workspace command with containers
With kas-container or custom docker mount-points, the
devtool-open-workspace command would fail to try to open the container's
workspace path on the host system.

Add the required logic to resolve the path to the host system. It's a
bit tricky if the container's mount points are different between the
sources and the build workspaces. We only support trivial cases for now
which work with kas-container.