CD: Automatically publish releases to the marketplace
Now that multiple target packages must be published, it's time to
automate the process, which will also properly document it.

The new workflow step relies on the VSCE_PAT secret, which has to be
configured and approved by the repository owner through the GitHub
settings page. It should contain a personal access token (PAT) with
permissions to publish packages to the marketplace.
This is documented here:
 - https://code.visualstudio.com/api/working-with-extensions/publishing-extension

The PAT have a limited lifetime, so they must be renewed periodically.

Of course I can't test this workflow untill a new release is published,
so it may require some tweaking.