Feat: Show complete history for variables on go-to-definition after the scan
The history exists in the form of comment above the variables in the scan results, thus they are extracted using the
existing functions that do this. The file path and locations are then extracted from the comments and later are used
in the onDefinition event. Seems like the duplicates will be automatically filtered by the vscode and each location
should be unique.