Fix: Provide references from definitions
When trying to "go to definition" on a single variable assignment,
VSCode would display "no references found" because of the setting
alternativeDefinitionCommand. We now provide references from
definitions, to work around this issue.

Definitions are references, but we don't provide full references
features. Other references are things like "${SRC_URI}".