Fix: can't process scan results if command is triggered by command palette
The command triggered by the command palette won't have the URI passed to the command handler. Thus, the value
of the chosen recipe is sent to the server to determine the URI by searching the scan results. Server will check
the uri and recipe in turn before proceeding while prioritizing URI over recipe.