Fix: store the results of the last scan to retain the values until next scan
The scan results are lost due to a new analyzedDocument is created on each document change. This is needed to get real-time
results from the tree-sitter. By storing the results of the last scan and read them on request could fix this.
Yet, the issue still happens on the comments or history field. Instead of duplicating the logics for this object, I plan to
refactor it so it should become part of the bitbakeSymbolInformation and be stored along the way.