Fix: return document links in order of appearance
Folder links Ctrl+Clicks still pointed to the default vscode resolver
while our "reveal" command appeared as a second option. It turns out
VSCode's algorithm to determine which one to show is by removing
duplicates in vscode/src/vs/editor/contrib/links/browser/getLinks.ts:_union

This algorithm only works if we provide the links in the same order. It
also works if we have gaps, but not if we are out of order. I changed
our way of building the links array to ensure that the links are in the
same order as they appear in the document.

This remove the duplicate default links for directories.