Fix: Running recipe scan and parse all recipes in a row cause the wrong text to show
These two tasks used the same boolean value to update the status, when they are triggered in a quick
succession, the onDidStartTask event is fired twice, when one of the them finishes, the boolean becomes false
and the status becomes non-parsing while the second one is still in process (which should have the status of
parsing). Thus, 1 boolean value is used for each process to keep the status correct.