Refactor: Allow relative paths in bitbake settings
Paths are no longer systematically resolved to absolute paths.
For the settings that concern starting bitbake, they'll be used as is
by the shell script. So we indicate in the UI that they should be
relative to the working directory instead of the workspace folder.

I've tested this did not produce regressions in the CI and a complex
docker with mountpoints.