Fix: Resolve container path in onDefinitionHandler
When using kas or other docker containers with mounted volumes, the
definitions history we get from bitbake -e contain paths inside the
container which do not correspond to the host paths.

The results of the bitbake -e scan live in the language server while the
resolver is in the client. I chose to make onDefinition request from the
server to the client because it happens only when requested so it
minimize the amount of data transfered.

The connection attribute is added in onDefinition because we can't
import the singleton from server.ts. It would crash in the Jest
environment.

This is not required for references since they come from our own parsing
rather than the bitbake -e recipe scan.