Fix: Use dynamic imports for node-pty
VSCode and the Jest native environment use conflicting
NODE_MODULE_VERSION. We have to use node-pty from VSCode
when running in the VSCode environment and the regular
node-pty when running in Jest.

Following the idea presented in:
https://github.com/microsoft/vscode/issues/658#issuecomment-982842847
We use dynamic imports to load the correct node-pty.

The types must still be imported normally at compile time.