Feat: Add settings and status bar for multiple configurations
A lot of users have requested to manage multiple build configurations.
It can be different build directories for different targets, varying kas
configurations, cqfd flavors, ...

We add a new setting `bitbake.buildConfigurations` which is an array of
all the bitbake invocation settings. If a setting is not defined in the
sub-configuration, the value from the main configuration will be used.
It's still possible to use just the main configuration.

The status bar will show the active configuration and allow to switch
between them.