Feat: Use settings to associate .conf and .inc files with the BitBake language
The extension gets activated when a .conf or .inc file is opened, but
they may also appear in non bitbake projects which is annoying for the
user. A bitbake error gets displayed and the parsing is wrong.

The alternative presented in this commit is to configure that
association after the activation. The cost is that we have to (once more)
modify settings.json. The tradeoff seems positive for me.