Fix: Handle request errors when using neovim
The latest completion changes make use of custom requests with the
client. When using the language server with nvim, CoC would crash the
server if an exception is thrown. We emit a log message instead.

Additionally, CoC starts the server with the `--stdio` argument, which
means any logs emitted will break the protocol. We add a setting to
disable logging in the recommended configuration.