CI: use the nodejs cache by not cleaning
The Github action setup-node has a built-in cache for node_modules,
which allows to save time by not having to reinstall all dependencies
every time. If we then clean and reinstall the dependencies, we lose
the benefit of the cache.

There is no need to clean the repository at the start of a pipeline
on non self-hosted runners.
Doc: https://github.com/marketplace/actions/clean-workspace