Chore: Use synchronize option to sync settings
Though this option is marked as deprecated, we can't fully embrach the recommended pull-model for syncing configuration
as the logger's logging level can't be synced in the same way like others. It needs to be updated immediately
instead of being fetched from the client when needed.
More info: https://code.visualstudio.com/api/language-extensions/language-server-extension-guide#explaining-the-language-server