Feat: Add a command to trigger global environment scan
This command is responsible for executing bitbake command `bitbake -e` and sending the output to the server for
further processing in order to provide necessary language features such as hover definition for conf files.
This commit only handles the command execution and the rest will be done in the following commits.