Fix: Adapt devtool status parsing for new Yocto releases
The current master branch also outputs the .bb location in the devtool
status output. It broke our parsing which could not detect the
workspaces.

The regex was updated to optionally match the .bb location. But we don't
capture it since we always run a complete scan in the end. This could
be way for future optimizations.