Refactor: Move scanResultVersion definition to BitbakeScanResult.ts
We use this variable to determine if the scan result data saved in the
workspace state is up to date. Since this should be modified whenever
the structure of the scan data changes, it makes sense to keep it in the
same file as the scan data structure.

Also added a reminder comment to increment the version number when the
structure changes.