Fix: task_deps parsing on newer Yocto versions
Two months ago, a commit was introduced that adds tasks with special
characters which broke our JSON parsing.

Since this variable is not extended, we might miss some tasks, but at
least we keep providing the feature. I could not identify a replacement
variable that would contained the expanded list of tasks.

We would need to run a `bitbake list-tasks` but it would be slower.