Scan: Add bitbake version to scan result
We'll use the bitbake version to make some features available or not
based on their support in the bitbake version.

The necessary ajustments were made to keep compiling.
The unknown version is '', making it easy to check if it was found or
not.
For the tests, rather than defining an empty version everywhere, I used
`as BitbakeScanResult` to mock only the attributes needed for each
test. This means less hassle next time we change the scan result
structure again.