Settings: Add disableDevtoolDebugBuild setting
By default, we'll turn on the new devtool debug build feature. This
allows accessing the debug symbols.

Because this is a new default behavior, we'll add a new setting to
disable this feature.

The necessary adjustments are made to compile the code by including a
boolean to the build settings.