Feature: Auto-rescan when changing to unknown buildConfiguration
Now that we can store individual scan results per buildConfiguration,
add the logic to trigger re-scans or UI updates when swapping.

Note that we don't invalidate the previous scan while doing the new one.
This allows the user to still navigate the recipes view, even though it
is no longer accurate for a few seconds. The scan is made visible in the
bottom right status bar.