Feat: Display active configuration in the recipes view title
As requested by a GitHub user. Now that the view is refreshed based on
the selected configuration, it will make it more obvious which config
is currently active. Note that it is also displayed in the bottom right
status bar.

If a re-scan is required, the title will be updated before the recipes
are displayed. This could be a bit misleading during the first config
swap.