Feat: support bitbake client debugging
This is a simple implementation of a command to generate a debug
configuration for the currently open BitBake client script. It creates a
launch.json entry with the necessary environment setup and allows
specifying arguments.

Example with debugging bitbake-getvar:
- Open the bitbake-getvar script.
- Press Ctrl + Shift + P and run "BitBake: Debug Current File"
- Enter arguments like `MACHINE`
- This will create a launch.json entry and start a debug session.
  If the debug sessions stops at with bb.utils.lock_timeout(self.rlock)
  just press "play" again to continue. Debugging bitbake with tinfoil
  is a bit tricky due to its architecture with the server and clients.
  But it is possible to set breakpoints in the client code.

Note about debugging the server: The server could be debugged by
attaching a debugger to the running server process. however, that is
more complex to set up. Not sure this is still possible due to the
double forking that happens when the server is started and the heartbeat
mechanism. So for now, this command focuses on debugging client scripts.

Maybe it would be better to have a dedicated debug configuration provider
class instead of generating launch.json directly. But this is more advanced.